+++ data/models/src/webui/mod.rs
@@ -1,0 +1,4 @@
+mod cluster;
+
+pub use self::cluster::TopClusterItem;
+pub use self::cluster::TopClusters;