+++ src/lib.rs
@@ -25,6 +25,7 @@ extern crate replicante_util_tracing;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
+extern crate serde_json;
 extern crate serde_yaml;
 
 #[macro_use]